+++ spec/units/gandi_v5/domain/availability/tax_spec.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+describe GandiV5::Domain::Availability::Tax do
+end